adiciona componentes e DTOs para gerenciamento de serviços e desempenho semanal

+++ dsin-cabeleleila-server/src/main/java/com/brunodias/dsin_cabeleleila_server/dtos/WeeklyPerformanceDTO.java
@@ -14,5 +14,5 @@ import java.math.BigDecimal;
 public class WeeklyPerformanceDTO {
     private int totalScheduledServices;
     private int totalCancelledServices;
-    private BigDecimal totalRevenue;
+    private Double totalRevenue;
 }
